% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexvolume.R
\name{intersect_vertices}
\alias{intersect_vertices}
\title{Compute the intersection vertices between two simplices. Wraps the
simplexintersection function from the SimplexIntersection.jl library
at https://github.com/kahaaga/SimplexIntersection.jl}
\usage{
intersect_vertices(S1, S2)
}
\arguments{
\item{S1}{The first simplex.}

\item{S2}{The second simplex.}
}
\description{
Compute the intersection vertices between two simplices. Wraps the
simplexintersection function from the SimplexIntersection.jl library
at https://github.com/kahaaga/SimplexIntersection.jl
}
