% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexvolume.R
\name{intersect_volume}
\alias{intersect_volume}
\title{Compute the intersection volume between two simplices. Wraps the
simplexintersection function from the SimplexIntersection.jl library
at https://github.com/kahaaga/SimplexIntersection.jl}
\usage{
intersect_volume(S1, S2)
}
\arguments{
\item{S1}{The first simplex.}

\item{S2}{The second simplex.}
}
\description{
Compute the intersection volume between two simplices. Wraps the
simplexintersection function from the SimplexIntersection.jl library
at https://github.com/kahaaga/SimplexIntersection.jl
}
