% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexvolume.R
\name{simplex_volume}
\alias{simplex_volume}
\title{Compute the intersection volume between two simplices. Wraps the
simplexintersection function from the SimplexIntersection.jl library
at https://github.com/kahaaga/SimplexIntersection.jl}
\usage{
simplex_volume(S1, S2)
}
\description{
Compute the intersection volume between two simplices. Wraps the
simplexintersection function from the SimplexIntersection.jl library
at https://github.com/kahaaga/SimplexIntersection.jl
}
